Remove compiler warning

+++ Core/Simulation/Simulation2D.cpp
@@ -47,9 +47,9 @@ std::unique_ptr<IComputation> Simulation2D::generateSingleThreadedComputation(si
                                                                               size_t n_elements)
 {
     assert(start < m_sim_elements.size() && start + n_elements <= m_sim_elements.size());
-    const auto& begin = m_sim_elements.begin() + start;
+    const auto& begin = m_sim_elements.begin() + static_cast<long>(start);
     return std::make_unique<DWBAComputation>(*sample(), m_options, m_progress, begin,
-                                             begin + n_elements);
+                                             begin + static_cast<long>(n_elements));
 }
 
 void Simulation2D::normalizeIntensity(size_t index, double beam_intensity)
